requirements.txt
fastapi
uvicorn
python-multipart
pydantic
sentence-transformers
faiss-cpu
requests
python-dotenv
PyPDF2
python-docx
numpy
